+++ src/invoices/InvoiceDetails/VoucherInformation/VoucherInformationContainer.js
@@ -41,7 +41,6 @@ const VoucherInformationContainer = ({ invoiceId, mutator, resources, location }
   useEffect(() => {
     mutator.voucher.reset();
     mutator.voucherLines.reset();
-
     mutator.voucher.GET().then(response => {
       const voucherId = get(response, '0.id');