+++ src/invoices/Voucher/VoucherEditForm/constants.js
@@ -1,0 +1,1 @@
+export const EDIT_VOUCHER_FORM = 'editVoucher';