+++ src/invoices/Voucher/VoucherEditForm/index.js
@@ -1,0 +1,1 @@
+export { default as VoucherEditContainer } from './VoucherEditContainer';