+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## (IN PROGRESS)
 
 ### Stories
+* [UINV-117](https://issues.folio.org/browse/UINV-117) Allow user to edit voucher information
 * [UINV-103](https://issues.folio.org/browse/UINV-103) Add voucher settings for a given Batch group
 * [UINV-118](https://issues.folio.org/browse/UINV-118) TECH-DEBT refactor Invoice list to not use SearchAndSort
 * [UINV-122](https://issues.folio.org/browse/UINV-122) Add "always show" preset adjustments to invoice when creating new invoice